Felddistanzen: 4D-Array zum Einmaligen ermitteln aller Distanzen
GameStateObserver: Abspeichern der Parameter als static Object GameState (innere Klasse), das bei jedem onGameStart() ueberschrieben werden muss

Heuristikfunktionen:
 - Heuristken nun von Typ float
 - sumDistanceToGhosts bugfix (je groesser der Wert, desto besser)
 - distanceToCloserGhosts: Heuristik fuer die Bewertung von Geisterentfernungen mit dem nahesten am hoechstbewertet und den weiter entfernten um eine weitere Kommastelle gerueckt

Knoten:
 - ghostsEaten verworfen, stattdessen:
  - powerpillTimer = 1 wenn Geist gefressen wurde (Bug, wenn geist suizid begeht) -> man wird nicht von respawnten geistern erwischt

MyAgent_PFinal:
 - Strategien:
  - 1: Mehrere Dots fressen (geht schlechter, warum auch immer)
  - 2: Ein Dot fressen
  - 3: Powerpille fressen
  - 4: Weglaufen mit Warten und ohne Risiko
   - Falls Warten zu gefaehrlich:
    - 5:  Weglaufen ohne Warten und mit Risiko

Zielfunktion: powerpillEaten

+++ src/de/fh/stud/Suchen/Suchkomponenten/InformedOpenList.java
@@ -6,12 +6,12 @@ import de.fh.stud.Suchen.Suche;
 import java.util.Comparator;
 import java.util.PriorityQueue;
 
-public class InformedOpenList extends OpenList{
+public class InformedOpenList extends OpenList {
     private final PriorityQueue<Knoten> openList;
 
     public InformedOpenList(Suche.SearchStrategy searchStrategy) {
         super(searchStrategy);
-        openList=new PriorityQueue<>(getInsertionCriteria(searchStrategy));
+        openList = new PriorityQueue<>(getInsertionCriteria(searchStrategy));
     }
 
     @Override
@@ -36,9 +36,9 @@ public class InformedOpenList extends OpenList{
 
     private static Comparator<Knoten> getInsertionCriteria(Suche.SearchStrategy strategy) {
         return switch (strategy) {
-            case GREEDY -> Comparator.comparingInt(Knoten::heuristicalValue);
+            case GREEDY -> Comparator.comparingDouble(Knoten::heuristicalValue);
             case UCS -> Comparator.comparingInt(Knoten::getCost);
-            case A_STAR -> Comparator.comparingInt(a -> a.getCost() + a.heuristicalValue());
+            case A_STAR -> Comparator.comparingDouble(a -> a.getCost() + a.heuristicalValue());
             default -> null;
         };
     }
