+++ src/de/fh/stud/interfaces/IHeuristicFunction.java
@@ -3,5 +3,5 @@ package de.fh.stud.interfaces;
 import de.fh.stud.Knoten;
 
 public interface IHeuristicFunction {
-	int calcHeuristic(Knoten node);
+    float calcHeuristic(Knoten node);
 }
